+++ Sawgle/TabViews/BookMarkViewController.swift
@@ -11,14 +11,10 @@ import UIKit
 class BookMarkViewController: UIViewController {
     override func loadView() {
         view = UIView()
-        view.backgroundColor = #colorLiteral(red: 0.9882352941, green: 0.9215686275, blue: 0.8235294118, alpha: 1)
+        view.backgroundColor = ColorList.pale
     }
 
     override func viewDidLoad() {
         navigationItem.title = "즐겨찾글"
-    }
-
-    override func viewWillAppear(_: Bool) {
-        print("BookMark")
     }
 }
